Second commit, about to start testing
Everything on my end should be in at this point. Kind of makes it easy
when you can just copy a shit ton of code from previous projects

Also added a very lame scoring system. There's a score integer that you
can rely on in the to print (may want to make scoring more interesting
later...and perhaps do a "star power" mode